Moving navbar of folder ⤴️
